+++ smog/const.py
@@ -6,3 +6,5 @@ DEFAULT_MEDIA_REPO = "~/media-repo"
 DEFAULT_MEDIA_DB = "~/media-db"
 
 DEFAULT_REL_PROC = "proc-media"
+
+SMOG_DB_NAME = "smog.db"