+++ smog_alembic/__init__.py
@@ -1,0 +1,3 @@
+from smog.file import FileStat
+
+# set env