+++ CHANGELOG.md
@@ -7,12 +7,12 @@ for open development tasks and limitations.
 
 ## version v0.0.2 - ???
 
-- cmd-line `xmp` sub-command for query xmp metadata
+- support of xmp metadata added
+  - support of various file types containing xmp metadata
+  - cmd-line `xmp` sub-command for query xmp metadata
 - sqlalchemy orm integration for media database
 - introduced pipe style processing of media items
 - enhanced cmd-line with `scan` sub-command
-- support of xmp metadata added
-  - support of various file types containing xmp metadata
 -