+++ CHANGELOG.md
@@ -10,7 +10,7 @@ for open development tasks and limitations.
 - support of xmp metadata added
   - support of various file types containing xmp metadata
   - cmd-line `xmp` sub-command for query xmp metadata
-- [`sqlalchemy`](https://github.com/sqlalchemy/sqlalchemy) orm integration for media database
+- [`sqlalchemy`](https://www.sqlalchemy.org/) orm integration for media database
 - introduced pipe style processing of media items
 - enhanced cmd-line with `scan` sub-command
 - cmd-line `find` sub-command